implementacion de deleteUser faltan detalles
